+++ lib/grpc.dart
@@ -21,8 +21,9 @@ export 'src/auth/auth.dart'
 
 export 'src/auth/auth_io.dart'
     show
-    ComputeEngineAuthenticator,
-    ServiceAccountAuthenticator;
+        applicationDefaultCredentialsAuthenticator,
+        ComputeEngineAuthenticator,
+        ServiceAccountAuthenticator;
 
 export 'src/client/call.dart' show CallOptions, ClientCall, MetadataProvider;
 export 'src/client/client.dart' show Client;