+++ lib/src/client/http2_connection.dart
@@ -53,6 +53,10 @@ class Http2ClientConnection implements connection.ClientConnection {
 
   /// Used for idle and reconnect timeout, depending on [_state].
   Timer _timer;
+
+  /// Used for making sure a single connection is not kept alive too long.
+  final Stopwatch _connectionLifeTimer = Stopwatch();
+
   Duration _currentReconnectDelay;
 
   final String host;
@@ -69,27 +73,49 @@ class Http2ClientConnection implements connection.ClientConnection {
 
   ConnectionState get state => _state;
 
-  Future<ClientTransportConnection> connectTransport() async {
-    var socket = await Socket.connect(host, port);
-    if (_state == ConnectionState.shutdown) {
-      socket.destroy();
-      // TODO(sigurdm): Throw something nicer...
-      throw 'Shutting down';
-    }
+  static const _estimatedRoundTripTime = const Duration(milliseconds: 20);
+
+  Future<Socket> _createSocket() async {
     final securityContext = credentials.securityContext;
-    if (securityContext != null) {
-      socket = await SecureSocket.secure(socket,
-          host: authority,
-          context: securityContext,
-          onBadCertificate: _validateBadCertificate);
-      if (_state == ConnectionState.shutdown) {
-        socket.destroy();
-        // TODO(sigurdm): Throw something nicer...
-        throw 'Shutting down';
+    if (securityContext == null) {
+      return Socket.connect(host, port);
+    } else {
+      if (options.credentials.authority == null) {
+        return SecureSocket.connect(host, port,
+            context: securityContext,
+            onBadCertificate: _validateBadCertificate);
+      } else {
+        // Todo(sigurdm): We want to pass supportedProtocols: ['h2']. http://dartbug.com/37950
+        return SecureSocket.secure(await Socket.connect(host, port),
+            // This is not really the host, but the authority to verify the TLC
+            // connection against.
+            //
+            // We don't use `this.authority` here, as that includes the port.
+            host: options.credentials.authority,
+            context: securityContext,
+            onBadCertificate: _validateBadCertificate);
       }
     }
-    socket.done.then((_) => _handleSocketClosed());
-    return ClientTransportConnection.viaSocket(socket);
+  }
+
+  Future<ClientTransportConnection> connectTransport() async {
+    final Socket socket = await _createSocket();
+    // Don't wait for io buffers to fill up before sending requests.
+    socket.setOption(SocketOption.tcpNoDelay, true);
+
+    final connection = ClientTransportConnection.viaSocket(socket);
+    socket.done.then((_) => _abandonConnection());
+
+    // Give the settings settings-frame a bit of time to arrive.
+    // TODO(sigurdm): This is a hack. The http2 package should expose a way of
+    // waiting for the settings frame to arrive.
+    await new Future.delayed(_estimatedRoundTripTime);
+
+    if (_state == ConnectionState.shutdown) {
+      socket.destroy();
+      throw _ShutdownException();
+    }
+    return connection;
   }
 
   void _connect() {
@@ -98,17 +124,44 @@ class Http2ClientConnection implements connection.ClientConnection {
       return;
     }
     _setState(ConnectionState.connecting);
-    connectTransport().then((transport) {
+    connectTransport().then((transport) async {
       _currentReconnectDelay = null;
       _transportConnection = transport;
+      _connectionLifeTimer
+        ..reset()
+        ..start();
       transport.onActiveStateChanged = _handleActiveStateChanged;
       _setState(ConnectionState.ready);
-      _pendingCalls.forEach(_startCall);
-      _pendingCalls.clear();
+
+      if (_hasPendingCalls()) {
+        // Take all pending calls out, and reschedule.
+        final pendingCalls = _pendingCalls.toList();
+        _pendingCalls.clear();
+        pendingCalls.forEach(dispatchCall);
+      }
     }).catchError(_handleConnectionFailure);
   }
 
+  /// Abandons the current connection if it is unhealthy or has been open for
+  /// too long.
+  ///
+  /// Assumes [_transportConnection] is not `null`.
+  void _refreshConnectionIfUnhealthy() {
+    final bool isHealthy = _transportConnection.isOpen;
+    final bool shouldRefresh =
+        _connectionLifeTimer.elapsed > options.connectionTimeout;
+    if (shouldRefresh) {
+      _transportConnection.finish();
+    }
+    if (!isHealthy || shouldRefresh) {
+      _abandonConnection();
+    }
+  }
+
   void dispatchCall(ClientCall call) {
+    if (_transportConnection != null) {
+      _refreshConnectionIfUnhealthy();
+    }
     switch (_state) {
       case ConnectionState.ready:
         _startCall(call);
@@ -210,9 +263,8 @@ class Http2ClientConnection implements connection.ClientConnection {
     }
     // TODO(jakobr): Log error.
     _cancelTimer();
-    _pendingCalls.forEach((call) => _failCall(call, error));
-    _pendingCalls.clear();
     _setState(ConnectionState.idle);
+    _connect();
   }
 
   void _handleReconnect() {
@@ -221,7 +273,7 @@ class Http2ClientConnection implements connection.ClientConnection {
     _connect();
   }
 
-  void _handleSocketClosed() {
+  void _abandonConnection() {
     _cancelTimer();
     _transportConnection = null;
 
@@ -275,3 +327,5 @@ class Http2ClientConnection implements connection.ClientConnection {
     return validator(certificate, authority);
   }
 }
+
+class _ShutdownException implements Exception {}