+++ CHANGELOG.md
@@ -1,3 +1,42 @@
+## 2.2.0+1
+
+* Relax `crypto` version dependency constraint from `^2.1.5` to `^2.1.4`.
+
+## 2.2.0
+
+* Added `applicationDefaultCredentialsAuthenticator` function for creating an
+  authenticator using [Application Default Credentials](https://cloud.google.com/docs/authentication/production).
+* Less latency by using the `tcpNoDelay` option for sockets.
+* Support grpc-web in a non-web setting.
+
+## 2.1.3
+
+* Fix bug in grpc-web when receiving an empty trailer.
+* Fix a state bug in the server.
+
+## 2.1.2
+
+* Fix bug introduced in 2.1.1 where the port would be added to the default authority when making a
+  secure connection.
+
+## 2.1.1
+
+* Fix bug introduced in 2.1.0 where an explicit `authority` would not be used when making a secure
+  connection.
+
+## 2.1.0
+
+* Do a health check of the http2-connection before making request.
+* Introduce `ChannelOptions.connectionLimit` the longest time a single connection is used for new
+  requests.
+* Use Tcp.nodelay to improve client call speed.
+* Use SecureSocket supportedProtocols to save a round trip when establishing a secure connection.
+* Allow passing http2 `ServerSettings` to `Server.serve`.
+
+## 2.0.3
+
+* GrpcError now implements Exception to indicate it can be reasonably handled.
+
 ## 2.0.2
 
 * Fix computation of the audience given to metadata providers to include the scheme.
@@ -8,7 +47,7 @@
 
 ## 2.0.0+1
 
-* Fix imports to ensure grpc-web.dart has no accidental transitive dependencies on dart:io.
+* Fix imports to ensure `grpc_web.dart` has no accidental transitive dependencies on dart:io.
 
 ## 2.0.0